requirements.txt
Flask==2.1.3
Werkzeug==2.0.3
requests==2.28.2
python-dateutil==2.8.2
python-dotenv==0.19.2
gunicorn==21.2.0